requirements.txt
fake_useragent
pillow
requests
requests_html